Messages with read & unread filter

+++ admin/orders.php
@@ -34,15 +34,12 @@ if (isset($_GET['delete_order'])) {
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta name="description"
-        content="EGYBOOK Bookstore is one of the leading book seller in Egypt. Boasting more than 7,000 Arabic and foreign titles and aiming to provide the best book shopping experience." />
+    <meta name="description" content="EGYBOOK Bookstore is one of the leading book seller in Egypt. Boasting more than 7,000 Arabic and foreign titles and aiming to provide the best book shopping experience." />
     <title>EGYBOOK</title>
     <!-- Tab Icon -->
     <link rel="shortcut icon" href="../img/logo.png" type="image/x-icon">
     <!-- Fontawesome CDN Link -->
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.1.1/css/all.min.css"
-        integrity="sha512-KfkfwYDsLkIlwQp6LFnl8zNdLGxu9YAA1QvwINks4PhcElQSvqcyVLLD9aMhXd13uQjoXtEKNosOWaZqXgel0g=="
-        crossorigin="anonymous" referrerpolicy="no-referrer" />
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.1.1/css/all.min.css" integrity="sha512-KfkfwYDsLkIlwQp6LFnl8zNdLGxu9YAA1QvwINks4PhcElQSvqcyVLLD9aMhXd13uQjoXtEKNosOWaZqXgel0g==" crossorigin="anonymous" referrerpolicy="no-referrer" />
     <!-- Custom CSS Style File -->
     <link rel="stylesheet" href="../css/admin_styles.css">
 </head>
@@ -63,28 +60,27 @@ if (isset($_GET['delete_order'])) {
 
             ?>
 
-            <div class="box">
-                <p>User ID: <span><?php echo $row['user_id'] ?></span></p>
-                <p>Placed On: <span><?php echo $row['placed_on'] ?></span></p>
-                <p>Name: <span><?php echo $row['name'] ?></span></p>
-                <p>number: <span><?php echo $row['number'] ?></span></p>
-                <p>email: <span><?php echo $row['email'] ?></span></p>
-                <p>address: <span><?php echo $row['address'] ?></span></p>
-                <p>total products: <span><?php echo $row['total_products'] ?></span></p>
-                <p>total price: <span>$<?php echo $row['total_price'] ?>/-</span></p>
-                <p>payment method: <span><?php echo $row['method'] ?></span></p>
-                <form action="" method="post">
-                    <input type="hidden" name="order_id" value="<?php echo $row['id'] ?>">
-                    <select name="update_payment">
-                        <option value="" selected disabled><?php echo $row['payment_status'] ?></option>
-                        <option value="pending">Pending</option>
-                        <option value="completed">Completed</option>
-                    </select>
-                    <input type="submit" value="Update" name="update_order" class="option-btn">
-                    <a href="orders.php?delete_order=<?php echo $row['id']; ?>" class="delete-btn"
-                        onclick="return confirm('Are you sure to delete this order?');">Delete</a>
-                </form>
-            </div>
+                    <div class="box">
+                        <p>User ID: <span><?php echo $row['user_id'] ?></span></p>
+                        <p>Placed On: <span><?php echo $row['placed_on'] ?></span></p>
+                        <p>Name: <span><?php echo $row['name'] ?></span></p>
+                        <p>number: <span><?php echo $row['number'] ?></span></p>
+                        <p>email: <span><?php echo $row['email'] ?></span></p>
+                        <p>address: <span><?php echo $row['address'] ?></span></p>
+                        <p>total products: <span><?php echo $row['total_products'] ?></span></p>
+                        <p>total price: <span>$<?php echo $row['total_price'] ?>/-</span></p>
+                        <p>payment method: <span><?php echo $row['method'] ?></span></p>
+                        <form action="" method="post">
+                            <input type="hidden" name="order_id" value="<?php echo $row['id'] ?>">
+                            <select name="update_payment">
+                                <option value="" selected disabled><?php echo $row['payment_status'] ?></option>
+                                <option value="pending">Pending</option>
+                                <option value="completed">Completed</option>
+                            </select>
+                            <input type="submit" value="Update" name="update_order" class="option-btn">
+                            <a href="orders.php?delete_order=<?php echo $row['id']; ?>" class="delete-btn" onclick="return confirm('Are you sure to delete this order?');">Delete</a>
+                        </form>
+                    </div>
 
             <?php
                 }
